+++ src/main/java/dev/matheuspereira/fluxcred/core/domain/model/Person.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -18,8 +19,8 @@ public class Person {
   private LocalDate birthDate;
   private String identifier;
   private IdentifierType identifierType;
-  private double minMonthlyPayment;
-  private double maxLoanAmount;
+  private BigDecimal minMonthlyPayment;
+  private BigDecimal maxLoanAmount;
   private LocalDateTime createdAt;
   private LocalDateTime updatedAt;
 }
