+++ src/main/java/dev/matheuspereira/fluxcred/core/infrastructure/entity/PersonEntity.java
@@ -18,6 +18,7 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -47,7 +48,7 @@ public class PersonEntity {
   private double minMonthlyPayment;
 
   @Column(nullable = false)
-  private double maxLoanAmount;
+  private BigDecimal maxLoanAmount;
 
   @Temporal(TemporalType.DATE)
   @Column(nullable = false)
